+++ skimage/filter/rank/_crank8.pyx
@@ -14,7 +14,7 @@ from skimage.filter.rank._core8 cimport _core8
 # -----------------------------------------------------------------
 
 
-cdef inline np.uint8_t kernel_autolevel(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_autolevel(Py_ssize_t * histo, float pop,
                                         np.uint8_t g, float p0, float p1,
                                         Py_ssize_t s0, Py_ssize_t s1):
 
@@ -38,7 +38,7 @@ cdef inline np.uint8_t kernel_autolevel(Py_ssize_t * histo, Py_ssize_t pop,
         return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_bottomhat(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_bottomhat(Py_ssize_t * histo, float pop,
                                         np.uint8_t g, float p0, float p1,
                                         Py_ssize_t s0, Py_ssize_t s1):
 
@@ -51,7 +51,7 @@ cdef inline np.uint8_t kernel_bottomhat(Py_ssize_t * histo, Py_ssize_t pop,
     return <np.uint8_t>(g - i)
 
 
-cdef inline np.uint8_t kernel_equalize(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_equalize(Py_ssize_t * histo, float pop,
                                        np.uint8_t g, float p0, float p1,
                                        Py_ssize_t s0, Py_ssize_t s1):
 
@@ -69,7 +69,7 @@ cdef inline np.uint8_t kernel_equalize(Py_ssize_t * histo, Py_ssize_t pop,
         return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_gradient(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_gradient(Py_ssize_t * histo, float pop,
                                        np.uint8_t g, float p0, float p1,
                                        Py_ssize_t s0, Py_ssize_t s1):
 
@@ -89,7 +89,7 @@ cdef inline np.uint8_t kernel_gradient(Py_ssize_t * histo, Py_ssize_t pop,
         return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_maximum(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_maximum(Py_ssize_t * histo, float pop,
                                       np.uint8_t g, float p0, float p1,
                                       Py_ssize_t s0, Py_ssize_t s1):
 
@@ -103,7 +103,7 @@ cdef inline np.uint8_t kernel_maximum(Py_ssize_t * histo, Py_ssize_t pop,
     return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_mean(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_mean(Py_ssize_t * histo, float pop,
                                    np.uint8_t g, float p0, float p1,
                                    Py_ssize_t s0, Py_ssize_t s1):
 
@@ -118,7 +118,7 @@ cdef inline np.uint8_t kernel_mean(Py_ssize_t * histo, Py_ssize_t pop,
         return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_meansubstraction(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_meansubstraction(Py_ssize_t * histo, float pop,
                                                np.uint8_t g, float p0, float p1,
                                                Py_ssize_t s0, Py_ssize_t s1):
 
@@ -133,7 +133,7 @@ cdef inline np.uint8_t kernel_meansubstraction(Py_ssize_t * histo, Py_ssize_t po
         return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_median(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_median(Py_ssize_t * histo, float pop,
                                      np.uint8_t g, float p0, float p1,
                                      Py_ssize_t s0, Py_ssize_t s1):
 
@@ -150,7 +150,7 @@ cdef inline np.uint8_t kernel_median(Py_ssize_t * histo, Py_ssize_t pop,
     return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_minimum(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_minimum(Py_ssize_t * histo, float pop,
                                       np.uint8_t g, float p0, float p1,
                                       Py_ssize_t s0, Py_ssize_t s1):
 
@@ -164,7 +164,7 @@ cdef inline np.uint8_t kernel_minimum(Py_ssize_t * histo, Py_ssize_t pop,
     return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_modal(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_modal(Py_ssize_t * histo, float pop,
                                     np.uint8_t g, float p0, float p1,
                                     Py_ssize_t s0, Py_ssize_t s1):
 
@@ -180,7 +180,7 @@ cdef inline np.uint8_t kernel_modal(Py_ssize_t * histo, Py_ssize_t pop,
     return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_morph_contr_enh(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_morph_contr_enh(Py_ssize_t * histo, float pop,
                                               np.uint8_t g, float p0, float p1,
                                               Py_ssize_t s0, Py_ssize_t s1):
 
@@ -203,14 +203,14 @@ cdef inline np.uint8_t kernel_morph_contr_enh(Py_ssize_t * histo, Py_ssize_t pop
         return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_pop(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_pop(Py_ssize_t * histo, float pop,
                                   np.uint8_t g, float p0, float p1,
                                   Py_ssize_t s0, Py_ssize_t s1):
 
     return <np.uint8_t>(pop)
 
 
-cdef inline np.uint8_t kernel_threshold(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_threshold(Py_ssize_t * histo, float pop,
                                         np.uint8_t g, float p0, float p1,
                                         Py_ssize_t s0, Py_ssize_t s1):
 
@@ -225,7 +225,7 @@ cdef inline np.uint8_t kernel_threshold(Py_ssize_t * histo, Py_ssize_t pop,
         return <np.uint8_t>(0)
 
 
-cdef inline np.uint8_t kernel_tophat(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_tophat(Py_ssize_t * histo, float pop,
                                      np.uint8_t g, float p0, float p1,
                                      Py_ssize_t s0, Py_ssize_t s1):
 
@@ -238,7 +238,7 @@ cdef inline np.uint8_t kernel_tophat(Py_ssize_t * histo, Py_ssize_t pop,
     return <np.uint8_t>(i - g)
 
 
-cdef inline np.uint8_t kernel_noise_filter(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_noise_filter(Py_ssize_t * histo, float pop,
                                            np.uint8_t g, float p0, float p1,
                                            Py_ssize_t s0, Py_ssize_t s1):
 
@@ -262,7 +262,7 @@ cdef inline np.uint8_t kernel_noise_filter(Py_ssize_t * histo, Py_ssize_t pop,
         return <np.uint8_t>min_i
 
 
-cdef inline np.uint8_t kernel_entropy(Py_ssize_t * histo, Py_ssize_t pop,
+cdef inline np.uint8_t kernel_entropy(Py_ssize_t * histo, float pop,
                                       np.uint8_t g, float p0, float p1,
                                       Py_ssize_t s0, Py_ssize_t s1):
     cdef Py_ssize_t i
@@ -278,7 +278,7 @@ cdef inline np.uint8_t kernel_entropy(Py_ssize_t * histo, Py_ssize_t pop,
     return <np.uint8_t>e*10
 
 
-cdef inline np.uint8_t kernel_otsu(Py_ssize_t * histo, Py_ssize_t pop, np.uint8_t g,
+cdef inline np.uint8_t kernel_otsu(Py_ssize_t * histo, float pop, np.uint8_t g,
                                    float p0, float p1, Py_ssize_t s0,
                                    Py_ssize_t s1):
     cdef Py_ssize_t i
